File created for Phat Smoke
Used to whip up a SVG format Nutritional Info card - made to fill current requirements. Enter details and export as SVG for use on products.

If I was to expand this I would likely make the document more dynamic with a standard template like this one or start from scratch and add lines manually.